korean_news_translator/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
html2text>=2020.1.16
